Added the ability to supply multiple passwords.
The choice was made to add support for multiple passwords instead of multiple credentials to remain backwards compatible. If we would add multiple usernames as well, it will would give the strange and error prone situation where name and password are matched based on location.

If there is a need to supply separate credentials we should make a non-backwards compatible change to properly add it.

Fixes #3